Test register() function directly instead of cmd_register
